Assignment  42 completed, Finally

+++ app/controllers/comments_controller.rb
@@ -5,26 +5,21 @@ class CommentsController < ApplicationController
 def create
   comment = Comment.new(comment_params)
   comment.user = current_user
+
  if params[:post_id]
     @post = Post.find(params[:post_id])
-# comment = @post.comments.new(comment_params)
     @post.comments << comment
-
-# comment.user = current_user
 
     if comment.save
       flash[:notice] = "Comment for post saved successfully."
       redirect_to [@post.topic, @post]
     else
       flash[:alert] = "Comment failed to save."
-      redirect_to [@post.topic, @po
+      redirect_to [@post.topic, @post]
     end
 
   elsif params[:topic_id]
     @topic = Topic.find(params[:topic_id])
-#    comment = @topic.comments.new(comment_params)
-
-
     @topic.comments << comment
 
     if comment.save
@@ -37,19 +32,50 @@ def create
   end
 end
 
-   def destroy
-     @post = Post.find(params[:post_id])
-     comment = @post.comments.find(params[:id])
+def destroy
+  comment = Comment.find(params[:id])
 
-     if comment.destroy
-       flash[:notice] = "Comment was deleted."
-       redirect_to [@post.topic, @post]
-     else
-       flash[:alert] = "Comment couldn't be deleted. Try again."
-       redirect_to [@post.topic, @post]
-     end
-   end
+  if params[:post_id]
+    @post = Post.find(params[:post_id])
+    comment = @post.comments.find(params[:id])
 
+      if comment.destroy
+        flash[:notice] = "Comment in post was deleted."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@post.topic, @post]
+      end
+
+  elsif params[:topic_id]
+    @topic = Topic.find(params[:topic_id])
+    comment = @topic.comments.find(params[:id])
+
+
+    if comment.destroy
+      flash[:notice] = "Comment in topic was deleted."
+      redirect_to [@topic]
+    else
+      flash[:alert] = "Comment couldn't be deleted. Try again."
+      redirect_to [@topic]
+    end
+  end
+end
+
+=begin
+def destroy
+  @post = Post.find(params[:post_id])
+  comment = @post.comments.find(params[:id])
+
+  if comment.destroy
+    flash[:notice] = "Comment was deleted."
+    redirect_to [@post.topic, @post]
+  else
+    flash[:alert] = "Comment couldn't be deleted. Try again."
+    redirect_to [@post.topic, @post]
+  end
+end
+=end
    private
 
    def comment_params
